Nfe - Fix - merge / FTF

+++ nds-service-client-impl/src/main/java/br/com/abril/nds/service/impl/GrupoServiceImpl.java
@@ -1,11 +1,15 @@
 package br.com.abril.nds.service.impl;
 
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.Date;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
+import org.apache.commons.lang.time.DateUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -43,11 +47,13 @@ public class GrupoServiceImpl implements GrupoService {
 	@Autowired
 	private DistribuidorRepository distribuidorRepository;
 	
+	private static final DateFormat DATE_FORMAT =  new SimpleDateFormat("dd/MM/yyyy");
+	
 	@Override
 	@Transactional
-	public List<GrupoCotaDTO> obterTodosGrupos(String sortname, String sortorder) {
+	public List<GrupoCotaDTO> obterTodosGrupos(String sortname, String sortorder, boolean includeHistory) {
 		
-		List<GrupoCota> grupos = grupoRepository.obterGrupos(sortname, sortorder);
+		List<GrupoCota> grupos = grupoRepository.obterGruposAtivos(sortname, sortorder, includeHistory);
 		
 		List<GrupoCotaDTO> gruposDTO = new ArrayList<GrupoCotaDTO>();
 
@@ -64,6 +70,11 @@ public class GrupoServiceImpl implements GrupoService {
 			dto.setRecolhimento(dias.toString());
 			dto.setTipoCota(grupo.getTipoCota());
 			dto.setTipoGrupo(grupo.getTipoGrupo());
+			
+			dto.setDataInicioVigencia(DATE_FORMAT.format(grupo.getDataInicioVigencia()));
+			if(grupo.getDataFimVigencia()!=null){
+			    dto.setDataFimVigencia(DATE_FORMAT.format(grupo.getDataFimVigencia()));
+			}
 			
 			gruposDTO.add(dto);			
 		}
@@ -92,20 +103,26 @@ public class GrupoServiceImpl implements GrupoService {
 		return dias;
 	}
 
-	@Transactional
+	@Transactional(readOnly=true)
 	@Override
-	public Integer countTodosGrupos() {
-		
-		return grupoRepository.countTodosGrupos();
+	public Integer countTodosGrupos(Date dataOperacao) {
+	    
+		return grupoRepository.countTodosGrupos(dataOperacao);
 	}
 	
 	@Transactional
 	@Override
 	public void excluirGrupo(Long idGrupo) {
-		
+		final Date dataOperacao = this.distribuidorRepository.obterDataOperacaoDistribuidor();
 		GrupoCota grupo = grupoRepository.buscarPorId(idGrupo);
 		
-		grupoRepository.remover(grupo);
+		if (grupo.getDataInicioVigencia().after(dataOperacao)){
+		    grupoRepository.remover(grupo);
+		}else{
+		    grupo.setDataFimVigencia(dataOperacao);
+	        grupoRepository.saveOrUpdate(grupo);
+		}
+		
 	}
 
 	@Override
@@ -138,69 +155,79 @@ public class GrupoServiceImpl implements GrupoService {
 	@Transactional
 	public void salvarGrupoCotas(Long idGrupo, Set<Long> idCotas, String nome,
 			List<DiaSemana> diasSemana, TipoCaracteristicaSegmentacaoPDV tipoCota) {
+	    final Date dataOperacao = this.distribuidorRepository.obterDataOperacaoDistribuidor();
+	    final Date dataOperacaoFutura = DateUtils.addDays(dataOperacao, 1);
+	    
+		this.validarNomeGrupo(nome, idGrupo, dataOperacaoFutura);
 		
-		this.validarNomeGrupo(nome, idGrupo);
+		Set<Cota> cotas = this.validarCotaJaPertenceGrupo(idGrupo, idCotas, null, dataOperacaoFutura);
 		
-		Set<Cota> cotas = this.validarCotaJaPertenceGrupo(idGrupo, idCotas, null);
-		
-		GrupoCota grupo;
-				
-		if(idGrupo == null) {
-			grupo = new GrupoCota();
-			grupo.setDataCadastro(this.distribuidorRepository.obterDataOperacaoDistribuidor());
-		} else {
-			grupo = grupoRepository.buscarPorId(idGrupo);
-		}
-		
-		grupo.setNome(nome);
-		grupo.setDiasRecolhimento(new HashSet<DiaSemana>(diasSemana));
-		grupo.setTipoGrupo(TipoGrupo.TIPO_COTA);
-		grupo.setTipoCota(tipoCota);
-		grupo.setCotas(cotas);
-		
-		grupoRepository.merge(grupo);
+		salvarGrupoCota(idGrupo, nome, diasSemana, TipoGrupo.TIPO_COTA, null, tipoCota, cotas,dataOperacao,dataOperacaoFutura);
 	}
 
 	@Override
 	@Transactional
 	public void salvarGrupoMunicipios(Long idGrupo, Set<String> municipios,
 			String nome, List<DiaSemana> diasSemana) {
+	    final Date dataOperacao = this.distribuidorRepository.obterDataOperacaoDistribuidor();
+	    final Date dataOperacaoFutura = DateUtils.addDays(dataOperacao, 1);
 		
-		this.validarNomeGrupo(nome, idGrupo);
+		this.validarNomeGrupo(nome, idGrupo, dataOperacaoFutura);
 		
-		this.validarCotaJaPertenceGrupo(idGrupo, null, municipios);
+		this.validarCotaJaPertenceGrupo(idGrupo, null, municipios,dataOperacaoFutura);
 		
-		GrupoCota grupo;
-		
-		if(idGrupo == null) {
-			grupo = new GrupoCota();
-			grupo.setDataCadastro(this.distribuidorRepository.obterDataOperacaoDistribuidor());
-		} else {
-			grupo = grupoRepository.buscarPorId(idGrupo);
-		}
-		
-		grupo.setNome(nome);
-		grupo.setDiasRecolhimento(new HashSet<DiaSemana>(diasSemana));
-		grupo.setTipoGrupo(TipoGrupo.MUNICIPIO);
-		grupo.setMunicipios(municipios);
-		
-		grupoRepository.merge(grupo);
+		salvarGrupoCota(idGrupo, nome, diasSemana, TipoGrupo.MUNICIPIO, municipios, null, null,dataOperacao,dataOperacaoFutura);
 	}
 
-	private void validarNomeGrupo(String nome, Long idGrupo) {
+    /**
+     * @param idGrupo
+     * @param municipios
+     * @param nome
+     * @param diasSemana
+     */
+    private void salvarGrupoCota(Long idGrupo, String nome, List<DiaSemana> diasSemana, TipoGrupo tipoGrupo, Set<String> municipios, TipoCaracteristicaSegmentacaoPDV tipoCota, Set<Cota> cotas, Date dataOperacao,Date dataOperacaoFutura) {
+       
+		
+		GrupoCota grupoNovo = null;
+		if(idGrupo != null) {
+		   final GrupoCota grupoAntigo = grupoRepository.buscarPorId(idGrupo);
+			if (grupoAntigo.getDataInicioVigencia().after(dataOperacao)){
+                grupoNovo = grupoAntigo;
+			} else {
+			    grupoAntigo.setDataFimVigencia(dataOperacao);
+			    grupoRepository.saveOrUpdate(grupoAntigo);
+            }
+			
+		}
+		
+		if(grupoNovo == null ){
+		    grupoNovo = new GrupoCota();
+	        grupoNovo.setDataInicioVigencia( dataOperacaoFutura);
+		}
+		grupoNovo.setNome(nome);
+		grupoNovo.setDiasRecolhimento(new HashSet<DiaSemana>(diasSemana));
+		grupoNovo.setTipoGrupo(tipoGrupo);
+		grupoNovo.setMunicipios(municipios);
+		grupoNovo.setTipoCota(tipoCota);
+		grupoNovo.setCotas(cotas);
+		
+		grupoRepository.merge(grupoNovo);
+    }
+
+	private void validarNomeGrupo(String nome, Long idGrupo, Date dataOperacao) {
 		
 		if (nome == null || nome.isEmpty()){
 			
 			throw new ValidacaoException(TipoMensagem.WARNING, "Informe um nome válido para o grupo!");
 		}
 		
-		if (this.grupoRepository.existeGrupoCota(nome, idGrupo)) {
+		if (this.grupoRepository.existeGrupoCota(nome, idGrupo, dataOperacao)) {
 			
 			throw new ValidacaoException(TipoMensagem.WARNING, "Já existe um grupo cadastrado com o nome: " + nome + "!");
 		}
 	}
 	
-	private Set<Cota> validarCotaJaPertenceGrupo(Long idGrupo, Set<Long> idCotas, Set<String> municipios){
+	private Set<Cota> validarCotaJaPertenceGrupo(Long idGrupo, Set<Long> idCotas, Set<String> municipios, Date dataOperacao){
 		
 		List<String> msgs = new ArrayList<>();
 		
@@ -213,14 +240,14 @@ public class GrupoServiceImpl implements GrupoService {
 				
 				Cota cota = cotaRepository.buscarPorId(id);
 				
-				String nomeGrupo = grupoRepository.obterNomeGrupoPorCota(id, idGrupo);
+				String nomeGrupo = grupoRepository.obterNomeGrupoPorCota(id, idGrupo, dataOperacao);
 				
 				if (nomeGrupo == null){
 					
 					if (cota.getEnderecoPrincipal() != null){
 						
 						nomeGrupo = this.grupoRepository.obterNomeGrupoPorMunicipio(
-							cota.getEnderecoPrincipal().getEndereco().getCidade(), null);
+							cota.getEnderecoPrincipal().getEndereco().getCidade(), null, dataOperacao);
 					}
 				}
 				
@@ -235,7 +262,7 @@ public class GrupoServiceImpl implements GrupoService {
 			
 			for (String municipio : municipios){
 				
-				String nomeGrupo = this.grupoRepository.obterNomeGrupoPorMunicipio(municipio, idGrupo);
+				String nomeGrupo = this.grupoRepository.obterNomeGrupoPorMunicipio(municipio, idGrupo, dataOperacao);
 				
 				if (nomeGrupo == null){
 					
@@ -243,7 +270,7 @@ public class GrupoServiceImpl implements GrupoService {
 					
 					for (Long idCota : cotasNoMunicipio){
 						
-						nomeGrupo = this.grupoRepository.obterNomeGrupoPorCota(idCota, idGrupo);
+						nomeGrupo = this.grupoRepository.obterNomeGrupoPorCota(idCota, idGrupo, dataOperacao);
 						
 						if(nomeGrupo != null){
 							msgs.add("Cota " + this.cotaRepository.buscarNumeroCotaPorId(idCota) +
